backend/ai/requirements.txt
alembic==1.17.2
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
deprecation==2.1.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gotrue==2.9.1
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.14.0
h2==4.3.0
hpack==4.1.0
httpcore==0.17.3
httplib2==0.31.0
httpx==0.24.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy==2.3.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
postgrest==0.13.2
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
realtime==1.0.6
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
storage3==0.7.7
StrEnum==0.4.15
supabase==2.3.0
supafunc==0.3.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
websockets==12.0
Werkzeug==3.1.4
